+++ src/Promitor.Tests.Integration/Clients/AgentClient.cs
@@ -54,7 +54,7 @@ namespace Promitor.Tests.Integration.Clients
             var context = new Dictionary<string, object>();
             try
             {
-                var rawResponse = await response.Content.ReadAsStringAsync();
+                await response.Content.ReadAsStringAsync();
                 // TODO: Uncomment for troubleshooting
                 // context.Add("Body", rawResponse);
             }
