+++ src/Promitor.Tests.Integration/IntegrationTest.cs
@@ -1,5 +1,6 @@
 ﻿using Arcus.Testing.Logging;
 using Microsoft.Extensions.Configuration;
+using Promitor.Tests.Integration.Infrastructure;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -16,11 +17,7 @@ namespace Promitor.Tests.Integration
             Logger = new XunitTestLogger(testOutput);
 
             // The appsettings.local.json allows users to override (gitignored) settings locally for testing purposes
-            Configuration = new ConfigurationBuilder()
-                .AddJsonFile(path: "appsettings.json")
-                .AddJsonFile(path: "appsettings.local.json", optional: true)
-                .AddEnvironmentVariables()
-                .Build();
+            Configuration = ConfigurationFactory.Create();
         }
     }
 }
